chore: splits client config into separate files
