+++ packages/payload/src/exports/config.ts
@@ -2,10 +2,13 @@ export { buildConfig } from '../config/build.js'
 export * from '../config/types.js'
 
 export type { FieldTypes } from '../admin/forms/FieldTypes.js'
+export { createClientCollectionConfig } from '../collections/config/client.js'
 export { defaults as collectionDefaults } from '../collections/config/defaults.js'
-export { createClientConfig } from '../config/createClientConfig.js'
+export { createClientConfig } from '../config/client.js'
 export { defaults } from '../config/defaults.js'
 export { sanitizeConfig } from '../config/sanitize.js'
 export { baseBlockFields } from '../fields/baseFields/baseBlockFields.js'
 export { baseIDField } from '../fields/baseFields/baseIDField.js'
+export { createClientFieldConfig } from '../fields/config/client.js'
 export { sanitizeFields } from '../fields/config/sanitize.js'
+export { createClientGlobalConfig } from '../globals/config/client.js'
