+++ packages/payload/src/exports/types.ts
@@ -15,6 +15,8 @@ export type {
   TabsPreferences,
 } from '../preferences/types.js'
 
+export type { ClientCollectionConfig } from './../collections/config/client.js'
+
 export type {
   AfterChangeHook as CollectionAfterChangeHook,
   AfterDeleteHook as CollectionAfterDeleteHook,
@@ -36,17 +38,9 @@ export type {
   TypeWithTimestamps,
 } from './../collections/config/types.js'
 
-export type {
-  Access,
-  AccessArgs,
-  ClientCollectionConfig,
-  ClientConfig,
-  ClientConfigField,
-  ClientGlobalConfig,
-  EditViewComponent,
-  SanitizedConfig,
-} from './../config/types.js'
-
+export type { ClientConfig } from './../config/client.js'
+export type { Access, AccessArgs, EditViewComponent, SanitizedConfig } from './../config/types.js'
+export type { ClientFieldConfig } from './../fields/config/client.js'
 export type {
   ArrayField,
   Block,
@@ -120,6 +114,8 @@ export {
   tabHasName,
   valueIsValueWithRelation,
 } from './../fields/config/types.js'
+
+export type { ClientGlobalConfig } from './../globals/config/client.js'
 
 export type {
   AfterChangeHook as GlobalAfterChangeHook,
